Office bugfixes

+++ src/TD.Web/TD.Web.Admin/TD.Web.Admin.Contracts/Requests/Products/ProductsGetMultipleRequest.cs
@@ -5,7 +5,7 @@ namespace TD.Web.Admin.Contracts.Requests.Products;
 public class ProductsGetMultipleRequest
 {
     public ProductClassification[]? Classification { get; set; }
-    public int[]? Groups { get; set; }
+    public long[]? Groups { get; set; }
     public string? SearchFilter { get; set; }
     public long[]? Id { get; set; }
     public ProductStatus[]? Status { get; set; }
